refactoring-and-getting-rid-of-the-odd-folders-making-project-one-branch-upper
