+++ static/restaurant/css/style.css
