+++ static/restaurant/js/scripts.js
