+++ templates/dish_confirm_delete.html
