+++ urls.py
@@ -14,7 +14,7 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.urls import path
-from restaurant_kitchen_mate import views
+import views
 
 urlpatterns = [
     path("", views.index, name="index"),
